moved FeaturesSearchRequestBase and FeaturesSearchResponseBase to the parent Request/Response folders
