+++ source/WesternStatesWater.WaDE.Contracts.Api/Requests/FeaturesSearchRequestBase.cs
@@ -1,6 +1,6 @@
 using WesternStatesWater.Shared.DataContracts;
 
-namespace WesternStatesWater.WaDE.Contracts.Api.Requests.V2;
+namespace WesternStatesWater.WaDE.Contracts.Api.Requests;
 
 public abstract class FeaturesSearchRequestBase : RequestBase
 {
